fix(services): align Apps governance endpoint files
Use plural route nouns for governance document and property endpoint filenames, and keep review-driven draft conversion non-blocking when GitHub rejects the workflow token mutation.
