+++ .github/workflows/pr-review-dismissed.yml
@@ -14,6 +14,8 @@ jobs:
       - uses: actions/checkout@v7
       - name: Set pull request to draft
         if: ${{ github.event.review.state == 'changes_requested' }}
+        # GitHub App tokens can reject the draft-conversion mutation even with pull-request write access.
+        continue-on-error: true
         run: gh pr ready ${{ github.event.pull_request.number }} --undo
         env:
           GH_TOKEN: ${{ github.token }}
